fix syntax tests
